+++ back/src/api/dbworks/def/def.go
@@ -1,0 +1,12 @@
+package def
+
+const (
+	DBAdmin   = "mysql"
+	User      = "tocchy"
+	Pass      = "*******"
+	Protocol  = "tcp(127.0.0.1:3306)"
+	Database  = "daily_todo"
+	Char      = "utf8mb4"
+	ParseTime = "True"
+	Loc       = "Local"
+)
