+++ back/src/api/dbworks/model/todo.go
@@ -1,0 +1,11 @@
+package model
+
+import "time"
+
+type ToDoList struct {
+	ID        int       `gorm:"column:id"`
+	UserID    int       `gorm:"column:user_id"`
+	Content   string    `gorm:"column:content"`
+	CreatedAt time.Time `gorm:"column:created_at"`
+	UpdatedAt time.Time `gorm:"column:updated_at"`
+}
