+++ back/src/api/dbworks/todo.go
@@ -1,0 +1,15 @@
+package dbworks
+
+import (
+	"./model"
+
+	"gorm.io/gorm"
+)
+
+func GetAll(db *gorm.DB) []model.ToDoList {
+
+	var todo []model.ToDoList
+	db.Find(&todo)
+
+	return todo
+}
